requirements/requirements.txt
kafka-python==2.0.2
pydantic==1.7.2
dependency-injector==4.32.2
orjson==3.5.2
structlog==20.2.0
colorama==0.4.4
